Edit csv generation, manage files
